Relocate CLI byte formatting helpers
Move formatByteSize/formatBytes from src/shared/utils/number.js to src/cli/lib/byte-format.js so the node-specific CLI utilities live with the consumers.
