+++ src/shared/tests/number-utils.test.js
@@ -1,11 +1,7 @@
 import assert from "node:assert/strict";
 import { describe, it } from "node:test";
 
-import {
-    formatByteSize,
-    isFiniteNumber,
-    toNormalizedInteger
-} from "../number-utils.js";
+import { isFiniteNumber, toNormalizedInteger } from "../number-utils.js";
 
 describe("number-utils", () => {
     describe("isFiniteNumber", () => {
@@ -22,33 +18,6 @@ describe("number-utils", () => {
             assert.equal(isFiniteNumber(Number.NaN), false);
             assert.equal(isFiniteNumber(Infinity), false);
             assert.equal(isFiniteNumber(-Infinity), false);
-        });
-    });
-
-    describe("formatByteSize", () => {
-        it("formats byte counts with default options", () => {
-            assert.equal(formatByteSize(0), "0B");
-            assert.equal(formatByteSize(512), "512B");
-            assert.equal(formatByteSize(2048), "2.0KB");
-        });
-
-        it("supports custom separators and precision", () => {
-            assert.equal(
-                formatByteSize(512, {
-                    decimals: 2,
-                    decimalsForBytes: 2,
-                    separator: " "
-                }),
-                "512.00 B"
-            );
-            assert.equal(
-                formatByteSize(5 * 1024 * 1024, {
-                    decimals: 2,
-                    separator: " ",
-                    trimTrailingZeros: true
-                }),
-                "5 MB"
-            );
         });
     });
